Aggregation structure

Create a price index aggregation structure from a hierarchical classification and aggregation weights.

Usage

aggregation_structure(x, w)

# S3 method for pias
weights(object, ea_only = FALSE, na.rm = FALSE, ...)

# S3 method for pias
update(object, index, period = end(index), ...)

# S3 method for pias
as.matrix(x, ...)

# S3 method for pias
as.data.frame(x, ..., stringsAsFactors = FALSE)

expand_classification(class, width)

Arguments

x

A list of character vectors (or vectors to be coerced to character) that give the codes/labels for the aggregation weights for each level of the classification, ordered so that moving down the list goes down the hierarchy. The last vector gives the elemental aggregates, which should have no duplicates. All vectors should be the same length, without NAs, and there should be no duplicates across different levels of x. This is usually a collection of columns in a data frame with a row for each aggregation weight; expand_classification() can be used when the classification is represented as a string.

w

A numeric vector of aggregation weights for the elemental aggregates (i.e., the last vector in x). The default is to give each elemental aggregate the same weight.

object

A price index aggregation structure, as made by aggregation_structure().

ea_only

Should weights be returned for only the elemental aggregates? The default gives the weights for the entire aggregation structure.

na.rm

Should missing values be removed from w when aggregating the weights (i.e., when ea_only = FALSE)? By default, missing values are not removed.

index

An aggregate price index, as made by aggregate().

period

The time period used to price update the weights. The default uses the last period in object.

class

A character vector, or something that can be coerced into one, of codes/labels for a specific level in a classification (e.g., 5-digit COICOP, 5-digit NAICS, 4-digit SIC).

width

A numeric vector that gives the width of each digit in x. This cannot contain NAs. The default assumes each digit has a width of 1, as in the NAICS, NAPCS, and SIC classifications.

stringsAsFactors

See as.data.frame

.

...

Further arguments passed to or used by methods.

Value

aggregation_structure() returns a price index aggregation structure. This is an object of class pias, which has the following components.

child

A nested list that gives the positions of the immediate children for each node in each level of the aggregation structure above the terminal nodes.

parent

A list that gives the position of the immediate parent for each node of the aggregation structure below the initial nodes.

levels

A character vector that gives the levels of x.

eas

A character vector that gives the subset of levels that are elemental aggregates.

weights

A named vector giving the weight for each elemental aggregate.

height

The length of x.

weights() returns a list with a named vector of weights for each level in the aggregation structure, unless ea_only = TRUE; in this case the return value is just a named vector.

update() returns a copy of object with price-updated weights using the index values in index.

as.matrix() represents an aggregation structure as a matrix, such that multiplying with a (column) vector of elemental indexes gives the aggregated index.

as.data.frame() takes an aggregation structure and returns a data frame that could have generated it, with columns level1, level2, ..., ea, and weight.

expand_classification() returns a list with a entry for each level in x giving the "digits" that represent each level in the hierarchy.

Warning

The aggregation_structure() function does its best to check its arguments, but there should be no expectation that the result of aggregation_structure() will make any sense if x does not represent a nested hierarchy.

See also

Examples

# A simple example
#            1
#      |-----+-----|
#      11          12
#  |---+---|       |
#  111     112     121

x1 <- c("1",   "1",   "1")
x2 <- c("11",  "11",  "12")
x3 <- c("111", "112", "121")

aggregation_structure(list(x1, x2, x3))
#> [[1]]
#> [1] "1"
#> 
#> [[2]]
#> [1] "11" "12"
#> 
#> [[3]]
#> [1] "111" "112" "121"
#> 

# The aggregation structure can also be made by expanding 'x3'

expand_classification(x3)
#> [[1]]
#> [1] "1" "1" "1"
#> 
#> [[2]]
#> [1] "11" "11" "12"
#> 
#> [[3]]
#> [1] "111" "112" "121"
#> 

all.equal(aggregation_structure(list(x1, x2, x3)), 
          aggregation_structure(expand_classification(x3)))
#> [1] TRUE

# Unequal weights

aggregation_structure(list(x1, x2, x3), 1:3)
#> [[1]]
#> [1] "1"
#> 
#> [[2]]
#> [1] "11" "12"
#> 
#> [[3]]
#> [1] "111" "112" "121"
#> 
          
# Extract the weights

weights(aggregation_structure(list(x1, x2, x3)))
#> [[1]]
#> 1 
#> 3 
#> 
#> [[2]]
#> 11 12 
#>  2  1 
#> 
#> [[3]]
#> 111 112 121 
#>   1   1   1 
#> 

# Expanding more complex classifications
# ... if last 'digit' is either TA or TS

expand_classification(c("111TA", "112TA", "121TS"), width = c(1, 1, 1, 2))
#> [[1]]
#> [1] "1" "1" "1"
#> 
#> [[2]]
#> [1] "11" "11" "12"
#> 
#> [[3]]
#> [1] "111" "112" "121"
#> 
#> [[4]]
#> [1] "111TA" "112TA" "121TS"
#> 

# ... if first 'digit' is either 11 or 12

expand_classification(c("111", "112", "121"), width = c(2, 1))
#> [[1]]
#> [1] "11" "11" "12"
#> 
#> [[2]]
#> [1] "111" "112" "121"
#> 

# ...if there are delimiters in the classification (like COICOP)

expand_classification(c("01.1.1", "01.1.2", "01.2.1"), width = c(2, 2, 2))
#> [[1]]
#> [1] "01" "01" "01"
#> 
#> [[2]]
#> [1] "01.1" "01.1" "01.2"
#> 
#> [[3]]
#> [1] "01.1.1" "01.1.2" "01.2.1"
#>