Price index objects

There are several classes to represent price indexes.

All indexes inherit from the piar_index virtual class.

Period-over-period indexes that can be chained over time inherit from chainable_piar_index.

Fixed-base indexes inherit from direct_piar_index.

Aggregate price indexes that are the result of aggregating elemental indexes with an aggregation structure further inherit from aggregate_piar_index.

Details

The piar_index object is a list-S3 class with the following components:

index

A list with an entry for each period in time that gives a vector of index values for each level in levels.

contrib

A list with an entry for each period in time, which itself contains a list with an entry for each level in levels with a named vector that gives the additive contribution for each price relative.

levels

A character vector giving the levels of the index.

time

A character vector giving the time periods for the index.

The chainable_piar_index and direct_piar_index subclasses have the same structure as the piar_index class, but differ in the methods used to manipulate the indexes.

The aggregate_piar_index class further subclasses either chainable_piar_index or direct_piar_index, and adds the following components:

r

The order of the generalized mean used to aggregated the index (usually 1).

pias

A list containing the child, parent, eas, and height components of the aggregation structured used to aggregate the index.