Price indexes

Compute period-over-period or fixed-base (direct) elemental price indexes (with optional quote contributions), or coerce pre-computed index values into an index object. There are methods to aggregate these indexes with a price index aggregation structure or average them over subperiods, combine indexes together, extract useful information, and coerce into a tabular form.

Usage

# Elemental indexes
elemental_index(rel, ...)

# S3 method for default
elemental_index(rel, ...)

# S3 method for numeric
elemental_index(rel, period = gl(1, length(rel)), 
                ea = gl(1, length(rel)), w, contrib = FALSE, 
                chainable = TRUE, na.rm = FALSE, r = 0, ...)
                            
# Aggregate
# S3 method for ind
aggregate(x, pias, na.rm = FALSE, r = 1, ...)

# S3 method for agg_ind
vcov(object, repweights, mse = TRUE, ...)

# S3 method for ind
mean(x, w, window = 3, na.rm = FALSE, r = 1, ...)

# Combine
# S3 method for ind
merge(x, y, ...)

# S3 method for ind
stack(x, y, ...)

# S3 method for ind
unstack(x, ...)

# Extract
# S3 method for ind
[(x, i, j)

# S3 method for ind
[(x, i, j) <- value

# S3 method for ind
levels(x)

# S3 method for ind
time(x, ...)

# S3 method for ind
start(x, ...)

# S3 method for ind
end(x, ...)

# S3 method for ind
head(x, n = 6, ...)

# S3 method for ind
tail(x, n = 6, ...)

# Coerce
# S3 method for ind
as.matrix(x, ...)

# S3 method for ind
as.data.frame(x, ..., stringsAsFactors = FALSE)
                            
as_index(x, ...)

# S3 method for default
as_index(x, ...)

# S3 method for matrix
as_index(x, chainable = TRUE, ...)

# S3 method for data.frame
as_index(x, cols = 1:3, chainable = TRUE, ...)

is_index(x)

is_aggregate_index(x)

Arguments

rel

Period-over-period or fixed-base price relatives. Currently there is only a method for numeric vectors; these can be made with price_relative().

period

A factor, or something that can be coerced into one, giving the time period associated with each price relative in rel. The ordering of time periods follows of the levels of period, to agree with cut(). The default assumes that all price relatives belong to one time period.

ea

A factor, or something that can be coerced into one, giving the elemental aggregate associated with each price relative in rel. The default assumes that all price relatives belong to one elemental aggregate.

w

A numeric vector of weights for the price relatives in rel/index values in x. The default is equal weights.

contrib

Should quote contributions be calculated? The default does not calculate contributions.

chainable

Are the price relatives in rel period-over-period relatives for a chained calculation (the default)? This should be FALSE when rel are fixed-base relatives.

na.rm

Should missing values be removed? By default, missing values are not removed. Setting na.rm = TRUE is equivalent to overall mean imputation.

r

Order of the generalized mean to aggregate price relatives/index values. 0 for a geometric index (the default for making elemental indexes), 1 for an arithmetic index (the default for aggregating elemental indexes and averaging indexes over subperiods), or -1 for a harmonic index (usually for a Paasche index). Other values are possible; see generalized_mean() for details.

x, y

A price index, usually made by elemental_index() or aggregate().

pias

A price index aggregation structure. This can be made with aggregation_structure().

object

An aggregate price index, as made by aggregate().

repweights

A matrix, or something that can be coerced into one, of bootstrap replicate weights with a row for each elemental aggregate and a column for each replicate.

mse

Should covariances be centered off the value of the index in object (the default), or the mean of the replicates?

window

The size of the window used to average index values across subperiods. The default (3) turns a monthly index into into a quarterly one.

i, j, value

See Extract, with value being a numeric vector (or something that can coerced into one.)

n

See head.

stringsAsFactors

See as.data.frame.

cols

A vector giving the positions/names of the period, level, and value columns in x. The default assumes that the first column contains time periods, the second contains levels, and the third contains index values.

...

Further arguments passed to or used by methods.

Details

Elemental indexes

When supplied with a numeric vector, elemental_index() is a simple wrapper that applies generalized_mean(r) and contributions(r) (if contrib = TRUE) to rel and w grouped by ea and period. That is, for every combination of elemental aggregate and time period, elemental_index() calculates an index based on a generalized mean of order r and, optionally, quote contributions (using names for rel as product names). The default (r = 0 and no weights) makes Jevons elemental indexes. See chapter 8 (pp. 175--190) of the CPI manual (2020) for more detail about making elemental indexes, and chapter 5 of Balk (2008).

The default method simply coerces rel to a numeric prior to calling the method above.

The interpretation of the index depends on how the price relatives in rel are made. If these are period-over-period relatives, then the result is a collection of period-over-period elemental indexes; if these are fixed-base relatives, then the result is a collection of fixed-base (direct) elemental indexes. For the latter, chainable should be set to FALSE so that no subsequent methods assume that a chained calculation should be used.

By default, missing price relatives in rel will propagate throughout the index calculation. Ignoring missing values with na.rm = TRUE is the same as parental (or overall mean) imputation, and needs to be explicitly set in the call to elemental_index(). Explicit imputation of missing relatives, and especially imputation of missing prices, should be done prior to calling elemental_index().

Indexes based on nested generalized means, like the Fisher index (and superlative quadratic mean indexes more generally), can be calculated by supplying the appropriate weights with nested_transmute(); see the example below. It is important to note that there are several ways to make these weights, and this affects how quote contributions are calculated.

Aggregate

The aggregate() method aggregates elemental indexes by looping over each time period in x and

aggregates the elemental indexes with generalized_mean(r) for each level of pias;

aggregates quote contributions for each level of pias (if there are any);

price updates the weights in pias with factor_weights(r) (only for period-over-period elemental indexes, i.e., is_chainable_index(x) == TRUE).

The result is a collection of aggregated period-over-period indexes that can be chained together to get a fixed-base index when x are period-over-period elemental indexes. Otherwise, when x are fixed-base elemental indexes, the result is a collection of aggregated fixed-base (direct) indexes.

By default, missing elemental indexes will propagate when aggregating the index. Missing elemental indexes can be due to both missingness of these values in x, and the presence of elemental aggregates in pias that are not part of x. Setting na.rm = TRUE ignores missing values, and is equivalent to parental (or overall mean) imputation. As an aggregated price index generally cannot have missing values (for otherwise it can't be chained over time), any missing values for a level of pias are removed and recursively replaced by the value of its immediate parent.

In most cases aggregation is done with an arithmetic mean (the default), and this is detailed in chapter 8 (pp. 190--198) of the CPI manual (2020). Aggregating with a non-arithmetic mean follows the same steps, except that the elemental indexes are aggregated with a mean of a different order (e.g., harmonic for a Paasche index), and the method for price updating the weights is slightly different.

Aggregating quote contributions uses the method in chapter 9 of the CPI manual (equations 9.26 and 9.28) when aggregating with an arithmetic mean. With a non-arithmetic mean, arithmetic weights are constructed using transmute_weights(r, 1) in order to apply this method.

There may not be contributions for all prices relatives in an elemental aggregate if the elemental indexes are built from several sources (as with merge()). In this case the contribution for a price relative in the aggregated index will be correct, but the sum of all contributions will not equal the change in the value of the index. This can also happen when aggregating an already aggregated index in which missing index values have been imputed (i.e., when na.rm = TRUE).

The vcov() method is a simple wrapper to calculate the variance matrix for an aggregated index when bootstrap replicate weights are available for the elemental aggregates. This approach is usually applicable when elemental aggregates are sampled, and provides an estimator of the sampling variance of the price index. It ignores any sampling variance from the elemental indexes (which often use judgmental sampling), and ultimately depends on the method of generating replicate weights. It returns a matrix of variances with a row for each upper-level index and a column for each time period. (Chapters 3 and 4 of Selvanathan and Rao (1994), especially section 4.7, provide analytic variance estimators for some common price indexes that are applicable with simple random sampling.) Note that any missing elemental indexes need to be explicitly imputed prior to using this method, otherwise they will propagate throughout the variance calculation.

Indexes can be aggregated over subperiods by taking the (usually arithmetic) mean of index values for each level over consecutive windows of subperiods. The mean() method constructs a set of windows of length window, starting in the first period of the index, and takes the unweighted mean of each index value in these windows for each level of the index. The last window is discarded if it is incomplete, so that index values are always averaged over window periods. The names for the first time period in each window form the new names for the aggregated time periods. Note that quote contributions are discarded when aggregating over subperiods.

An optional vector of weights can be specified when aggregating index values over subperiods, which is often useful when aggregating a Paasche index; see section 4.3 of Balk (2008) for details. It is usually easiest to specify these weights as a matrix with a row for each index value in x and a column for each time period.

Combine

The merge() method combines two index objects with common time periods, merging together the index values and quote contributions for each time period in x and y. This is useful for building up an index when different elemental aggregates come from different sources of data, or use different index-number formulas.

The stack() method combines two index objects with common levels, stacking index values and quote contributions for each level in y after those in x. The unstack() method breaks up x into a list of indexes, one for each period in x. These methods can be used in a map-reduce to make an index with multiple aggregation structures (like a Paasche index).

It is not generally possible to merge aggregated indexes, as this would change the aggregation structure, so merging always returns an index of class ind. If at least one of x or y is an aggregate index then the result of stacking these indexes is also an aggregate index; otherwise, it is the same class as x.

Extract

The extraction method treats x as a matrix of index values with (named) rows for each level and columns for each period in x. Unlike a matrix, dimensions are never dropped as indexing x always returns an index object. This means that indexing with a matrix is not possible, and only a submatrix can be extracted. As x is not an atomic vector, indexing with a single index like x[1] is taken to be the same as x[1, ]. Note that indexing an aggregated index cannot generally preserve the aggregation structure if any levels are removed, and in this case the resulting index is not an aggregated index.

The replacement method similarly treats x as a matrix, and behaves the same as replacing values in a matrix (except that value is coerced to numeric). Note that replacing the values of an index will remove the corresponding quote contributions (if any).

The levels() method extracts the levels of an index, and the time() method extracts the time periods of the index. The start() and end() methods extract the first and last time period.

The head() and tail() methods act as if x is a matrix of index values, and by default extract the time series for the first/last six levels of x.

The summary() method summarizes x as a matrix of index values (i.e., the five-number summary for each period). If there are quote contributions, then these are also summarized as a matrix.

Coerce

The as.matrix() method turns an index into a matrix with a row for each level and a column for each period. The as.data.frame() method turns an index into a data frame with three columns: period, level, and value.

as_index() is useful to form an elemental index from pre-computed values. Numeric matrices are coerced into an elemental index object by treating each column as a separate time period, and each row as an elemental aggregate. Column names are used to denote time periods, and row names are used to denote elemental aggregates (so they must be unique). This essentially reverses calling as.matrix() on an index object. If a dimension is unnamed, then it is given a sequential label from 1 to the size of that dimension. The default method coerces x to a matrix prior to using the matrix method.

The data frame method for as_index() is best understood as reversing the effect of as.data.frame() on an index object. It constructs a matrix by taking the unique values of x[[cols[1]]] as columns and the unique values of x[[cols[2]]] as rows (in the order they appear). It then populates this matrix with the corresponding values in x[[cols[3]]], and uses the matrix method for as_index(). Note that the resulting index is therefore sensitive to the ordering of x.

Value

Most of these functions return index objects of class ind. These often behaves like a matrix with a row for each level of the index and a column for each time period, and have the following components.

index

A named list with an entry for each period that gives a named vector of index values for each level in ea.

contrib

A named list with an entry for each period, which itself contains a named list with an entry for each level in ea with a named vector that gives the additive contribution for each price relative. If contrib = FALSE, then each of these vectors is of length 0.

levels

The levels for ea.

time

The levels for period.

has_contrib

The value of contrib.

chainable

The value of chainable, usually TRUE.

aggregate() returns an aggregate index. This is an object of class agg_ind, inheriting from class ind, which has the following components.

index

A named list with an entry for each period in x that gives a named vector of index values for each level in pias.

contrib

A named list with an entry for each period, which itself contains a list with an entry for each level in pias with a named vector that gives the additive contribution for each price relative.

levels

The levels for pias.

time

The levels for period from x.

has_contrib

The value of has_contrib from x.

chainable

The value for chainable from x, usually TRUE.

r

The value for r, usually 1.

pias

A list containing the child, parent, eas, and height components of pias.

References

Balk, B. M. (2008). Price and Quantity Index Numbers. Cambridge University Press.

ILO, IMF, OECD, Eurostat, UN, and World Bank. (2020). Consumer Price Index Manual: Theory and Practice. International Monetary Fund.

Selvanathan, E. A., and Rao, D. S. P. (1994). Index Numbers: A Stochastic Approach. MacMillan.

Source

The vcov() method was influenced by a SAS routine by Justin Francis that was first ported to R by Ambuj Dewan, and subsequently rewritten by Steve Martin.

See also

Examples

prices <- data.frame(rel = 1:8, period = rep(1:2, each = 4), ea = rep(letters[1:2], 4))

# A two-level aggregation structure

pias <- aggregation_structure(list(c("top", "top", "top"), c("a", "b", "c")), 1:3)

# Calculate Jevons elemental indexes

(epr <- with(prices, elemental_index(rel, period, ea)))
#>          1        2
#> a 1.732051 5.916080
#> b 2.828427 6.928203

# Same as using lm() or tapply()

exp(coef(lm(log(rel) ~ ea:factor(period) - 1, prices)))
#> eaa:factor(period)1 eab:factor(period)1 eaa:factor(period)2 eab:factor(period)2 
#>            1.732051            2.828427            5.916080            6.928203 

with(prices, t(tapply(rel, list(period, ea), gpindex::geometric_mean, na.rm = TRUE)))
#>          1        2
#> a 1.732051 5.916080
#> b 2.828427 6.928203

# Extract the indexes like a matrix

epr["a", ]
#>          1       2
#> a 1.732051 5.91608

epr[, 2]
#>          2
#> a 5.916080
#> b 6.928203

epr[1, ] <- 1 # can be useful for doing specific imputations

# Aggregate (note the imputation for elemental index 'c')

(index <- aggregate(epr, pias, na.rm = TRUE))
#>            1        2
#> top 2.218951 6.037662
#> a   1.000000 1.000000
#> b   2.828427 6.928203
#> c   2.218951 6.037662

# Aggregation can equivalently be done as matrix multiplication

as.matrix(pias) %*% as.matrix(chain(index[letters[1:3]]))
#>            1        2
#> top 2.218951 13.39728

# Merge two indexes prior to aggregation

prices2 <- data.frame(rel = 1:8, period = rep(1:2, each = 4), ea = rep(letters[3:4], 4))
epr2 <- with(prices2, elemental_index(rel, period, ea))
aggregate(merge(epr, epr2), pias)
#>            1        2
#> top 1.975501 5.984362
#> a   1.000000 1.000000
#> b   2.828427 6.928203
#> c   1.732051 5.916080

# Stack two indexes prior to aggregation

prices3 <- data.frame(rel = 1:8, period = rep(3:4, each = 4), ea = rep(letters[1:2], 4))
epr3 <- with(prices3, elemental_index(rel, period, ea))
aggregate(stack(epr, epr3), pias)
#>            1        2        3        4
#> top       NA       NA       NA       NA
#> a   1.000000 1.000000 1.732051 5.916080
#> b   2.828427 6.928203 2.828427 6.928203
#> c         NA       NA       NA       NA

# Unstack does the reverse

all.equal(c(unstack(epr), unstack(epr3)), unstack(stack(epr, epr3)))
#> [1] TRUE

# Extract useful features of the index

head(index, 1)
#>            1        2
#> top 2.218951 6.037662
tail(index, 3)
#>          1        2
#> a 1.000000 1.000000
#> b 2.828427 6.928203
#> c 2.218951 6.037662
levels(index)
#> [1] "top" "a"   "b"   "c"  
time(index)
#> [1] "1" "2"
start(index)
#> [1] "1"
end(index)
#> [1] "2"

summary(index)
#> Indexes
#>        1               2        
#>  Min.   :1.000   Min.   :1.000  
#>  1st Qu.:1.914   1st Qu.:4.778  
#>  Median :2.219   Median :6.038  
#>  Mean   :2.067   Mean   :5.001  
#>  3rd Qu.:2.371   3rd Qu.:6.260  
#>  Max.   :2.828   Max.   :6.928  

# Turn the index into a data frame/matrix

as.data.frame(index)
#>   period level    value
#> 1      1   top 2.218951
#> 2      1     a 1.000000
#> 3      1     b 2.828427
#> 4      1     c 2.218951
#> 5      2   top 6.037662
#> 6      2     a 1.000000
#> 7      2     b 6.928203
#> 8      2     c 6.037662
as.matrix(index)
#>            1        2
#> top 2.218951 6.037662
#> a   1.000000 1.000000
#> b   2.828427 6.928203
#> c   2.218951 6.037662

all.equal(as_index(as.data.frame(epr)), epr)
#> [1] TRUE
all.equal(as_index(as.matrix(epr)), epr)
#> [1] TRUE

# Calculate a CSWD index (same as the Jevons in this example) 
# as an arithmetic index by constructing appropriate weights

library(gpindex)

# A general function to calculate weights to turn the geometric
# mean of the arithmetic and harmonic mean (i.e., Fisher mean)
# into an arithmetic mean

fw <- grouped(nested_transmute(0, c(1, -1), 1))

with(
    prices, 
    elemental_index(rel, period, ea, fw(rel, group = interaction(period, ea)), r = 1)
)
#>          1        2
#> a 1.732051 5.916080
#> b 2.828427 6.928203