Make a price index aggregation structure

Create a price index aggregation structure from a hierarchical classification and aggregation weights that can be used to aggregate elemental indexes.

Usage

aggregation_structure(x, w = NULL)

Arguments

x

A list of character vectors that give the codes/labels for each level of the classification, ordered so that moving down the list goes down the hierarchy. The last vector gives the elemental aggregates, which should have no duplicates. All vectors should be the same length, without NAs, and there should be no duplicates across different levels of x.

w

A numeric vector of aggregation weights for the elemental aggregates (i.e., the last vector in x). The default is to give each elemental aggregate the same weight.

Value

A price index aggregation structure. This is an object of class aggregation_structure, which has the following components.

child

A nested list that gives the positions of the immediate children for each node in each level of the aggregation structure above the terminal nodes.

parent

A list that gives the position of the immediate parent for each node of the aggregation structure below the initial nodes.

levels

A character vector that gives the levels of x.

eas

A character vector that gives the subset of levels that are elemental aggregates.

weights

A named vector giving the weight for each elemental aggregate.

height

The length of x.

Warning

The aggregation_structure() function does its best to check its arguments, but there should be no expectation that the result of aggregation_structure() will make any sense if x does not represent a nested hierarchy.

See also

Examples

# A simple aggregation structure
#            1
#      |-----+-----|
#      11          12
#  |---+---|       |
#  111     112     121
#  (1)     (3)     (4)

aggregation_weights <- data.frame(
  level1 = c("1",   "1",   "1"),
  level2 = c("11",  "11",  "12"),
  ea     = c("111", "112", "121"),
  weight = c(1,     3,     4)
)
                                  
pias <- aggregation_structure(
  aggregation_weights[1:3], w = aggregation_weights[[4]]
)

# The aggregation structure can also be made by expanding the
# elemental aggregates

all.equal(
  with(aggregation_weights,
       aggregation_structure(expand_classification(ea), weight)),
  pias
)
#> [1] TRUE