Coerce to a price index

Coerce pre-computed index values into an index object, and test if an object is a price index.

Usage

as_index(x, ...)

# S3 method for default
as_index(x, chainable = TRUE, ...)

# S3 method for matrix
as_index(x, chainable = TRUE, ...)

# S3 method for data.frame
as_index(x, cols = 1:3, chainable = TRUE, ...)

# S3 method for chainable_pindex
as_index(x, chainable = TRUE, ...)

# S3 method for direct_pindex
as_index(x, chainable = FALSE, ...)

is_index(x)

is_chainable_index(x)

is_direct_index(x)

is_aggregate_index(x)

Arguments

x

An object to test or coerce into a price index.

chainable

Are the index values in x period-over-period indexes, suitable for a chained calculation (the default)? This should be FALSE when x is a fixed-base (direct) index.

cols

A vector giving the positions/names of the period, level, and value columns in x. The default assumes that the first column contains time periods, the second contains levels, and the third contains index values.

...

Further arguments passed to or used by methods.

Details

Numeric matrices are coerced into an index object by treating each column as a separate time period, and each row as an elemental aggregate. Column names are used to denote time periods, and row names are used to denote elemental aggregates (so they must be unique). This essentially reverses calling as.matrix() on an index object. If a dimension is unnamed, then it is given a sequential label from 1 to the size of that dimension. The default method coerces x to a matrix prior to using the matrix method.

The data frame method for as_index() is best understood as reversing the effect of as.data.frame() on an index object. It constructs a matrix by taking the levels of x[[cols[1]]] as columns and the levels of x[[cols[2]]] as rows (coercing to a factor if necessary). It then populates this matrix with the corresponding values in x[[cols[3]]], and uses the matrix method for as_index().

If x is a period-over-period index then it is returned unchanged when chainable == TRUE and chained otherwise. Similarly, if x is a fixed-base index then it is returned unchanged when chainable == FALSE and unchain otherwise.

Value

as_index() returns a price index that inherits from pindex. If chainable == TRUE then this is a period-over-period price index that also inherits from chainable_pindex; otherwise, it is a fixed-base index that inherits from direct_pindex.

is_index() returns TRUE if x inherits from pindex.

is_chainable_index() returns TRUE if x inherits from chainable_pindex.

is_direct_index() returns TRUE if x inherits from direct_pindex.

is_aggregate_index() returns TRUE if x inherits from aggregate_pindex.

See also

Examples

prices <- data.frame(
  rel = 1:8,
  period = rep(1:2, each = 4),
  ea = rep(letters[1:2], 4)
)

pias <- aggregation_structure(
  list(c("top", "top", "top"), c("a", "b", "c")), 1:3
)

# Calculate period-over-period Jevons elemental indexes

epr <- with(prices, elemental_index(rel, period, ea))

# Aggregate as an arithmetic index

index <- aggregate(epr, pias, na.rm = TRUE)

is_index(epr)
#> [1] TRUE
is_chainable_index(epr)
#> [1] TRUE
is_direct_index(epr)
#> [1] FALSE
is_aggregate_index(epr)
#> [1] FALSE

is_index(index)
#> [1] TRUE
is_chainable_index(index)
#> [1] TRUE
is_direct_index(index)
#> [1] FALSE
is_aggregate_index(index)
#> [1] TRUE

all.equal(as_index(as.data.frame(epr)), epr)
#> [1] TRUE
all.equal(as_index(as.matrix(epr)), epr)
#> [1] TRUE