Coerce to an aggregation structure

Coerce an object into an aggregation structure object, and test if an object is an aggregation structure.

Usage

as_aggregation_structure(x, ...)

# S3 method for default
as_aggregation_structure(x, w = NULL, ...)

# S3 method for data.frame
as_aggregation_structure(x, ...)

# S3 method for matrix
as_aggregation_structure(x, ...)

# S3 method for aggregate_index
as_aggregation_structure(x, w = NULL, ...)

# S3 method for aggregation_structure
as_aggregation_structure(x, ...)

is_aggregation_structure(x)

Arguments

x

An object to test or coerce into an aggregation structure.

w

A numeric vector of aggregation weights for the elemental aggregates. The default is to give each elemental aggregate the same weight.

...

Further arguments passed to or used by methods.

Details

The default method attempts to coerce x into a list prior to calling aggregation_structure().

The data frame and matrix methods treat x as a table with a row for each elemental aggregate, a column of labels for each level in the aggregation structure, and a column of weights for the elemental aggregates.

The method for aggregate indexes reconstructs the aggregation structure used to generate the index (with optional weights).

Value

A price index aggregation structure, except for is_aggregate_index() that returns TRUE if x inherits from aggregation_structure.

See also

Examples

# A simple aggregation structure
#            1
#      |-----+-----|
#      11          12
#  |---+---|       |
#  111     112     121
#  (1)     (3)     (4)

aggregation_weights <- data.frame(
  level1 = c("1",   "1",   "1"),
  level2 = c("11",  "11",  "12"),
  ea     = c("111", "112", "121"),
  weight = c(1,     3,     4)
)
                                  
pias <- aggregation_structure(aggregation_weights[1:3],
                              w = aggregation_weights[[4]])
                              
all.equal(
  pias,
  as_aggregation_structure(aggregation_weights)
)
#> [1] TRUE

all.equal(
  pias,
  as_aggregation_structure(as.matrix(aggregation_weights))
)
#> [1] TRUE