Merge price indexes

Combine two price indexes with common time periods, merging together the index values and percent-change contributions for each time period.

This is useful for building up an index when different elemental aggregates come from different sources of data, or use different index-number formulas.

Usage

# S3 method for piar_index
merge(x, y, ...)

Arguments

x

A price index, as made by, e.g., elemental_index().

y

A price index, or something that can coerced into one. If x is a period-over-period index then y is coerce into a chainable index; otherwise, y is coerced into a direct index.

...

Further arguments passed to or used by methods.

Value

A price index that inherits from chainable_piar_index if x is a period-over-period index, or direct_piar_index if x is a fixed-base index. It is not generally possible to merge aggregated indexes, as this would change the aggregation structure, so merging does not return an aggregated index.

See also

Examples

prices <- data.frame(
  rel = 1:8,
  period = rep(1:2, each = 4),
  ea = rep(letters[1:2], 4)
)

prices2 <- data.frame(
  rel = 1:8,
  period = rep(1:2, each = 4),
  ea = rep(letters[3:4], 4)
)

epr <- with(prices, elemental_index(rel, period, ea))

epr2 <- with(prices2, elemental_index(rel, period, ea))

merge(epr, epr2)
#> Period-over-period price index for 4 levels over 2 time periods 
#>          1        2
#> a 1.732051 5.916080
#> b 2.828427 6.928203
#> c 1.732051 5.916080
#> d 2.828427 6.928203