piar: Price Index Aggregation

Most price indexes are made with a two-step procedure, where period-over-period elemental indexes are first calculated for a collection of elemental aggregates at each point in time, and then aggregated according to a price index aggregation structure. These indexes can then be chained together to form a time series that gives the evolution of prices with respect to a fixed base period. This package contains a collection of functions that revolve around this work flow, making it easy to build standard price indexes, and implement the methods described by Balk (2008, doi:10.1017/CBO9780511720758 ), von der Lippe (2007, doi:10.3726/978-3-653-01120-3 ), and the CPI manual (2020, doi:10.5089/9781484354841.069 ) for bilateral price indexes.

See also

Author

Maintainer: Steve Martin marberts@protonmail.com (ORCID) [copyright holder]